refactored image controller and service
