+++ web/.gitignore
@@ -114,3 +114,7 @@ npm-debug.log
 # sample images and output
 images/sample/
 images/out/
+
+*.pgm
+*.png
+*.yml
